backend/requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.29,<1
pydantic>=2,<3
pydantic-settings>=2,<3
httpx>=0.27,<1
boto3>=1.34,<2
google-cloud-texttospeech>=2.16,<3
edge-tts>=7.2,<8
markitdown>=0.1,<1
python-multipart>=0.0.9,<1
redis[hiredis]>=5.0,<6
arq>=0.25,<1
openai>=1.12,<2
pytesseract>=0.3,<1
pyyaml>=6,<7
readability-lxml>=0.8,<1
Pillow>=10,<11
